feat: change HistoryCard to dumb component type
