+++ src/components/HistoryCard.tsx
@@ -1,35 +1,16 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import React from 'react';
-import { useNavigate } from 'react-router-dom';
 import PropTypes, { InferProps } from 'prop-types';
-import { useAppDispatch } from '../redux/hooks';
-import { addHistory } from '../redux/userSlice';
 import { objMaker } from '../tools/objMaker';
-import { makeUrl } from '../tools/urlMaker';
-import UserDataHandlerToLS from '../utils/userDataWriter';
 import styles from './HistoryCard.module.css';
 
 function HistoryCard(props: HistoryCardProps) {
-  const { item, refreshParent } = props;
-
-  const { email } = UserDataHandlerToLS.getCurrentUser();
-  const navigate = useNavigate();
-  const dispatch = useAppDispatch();
-
+  const { item, refreshParent, searchHistory } = props;
   const [keyArr, valArr] = objMaker(item);
   const keyHandler = (key: string) => {
     if (key === 'q') return 'Your search';
     if (key.includes('Type')) return key.replace('T', ' t');
     return key;
-  };
-  const handleSearch = () => {
-    const query = makeUrl(item);
-    dispatch(addHistory(query));
-    navigate('/search');
-  };
-  const handleDelete = () => {
-    UserDataHandlerToLS.removeFromHistory(email, item);
-    refreshParent(item);
   };
   return (
     <div className={styles.searchQueryBlock}>
@@ -37,8 +18,8 @@ function HistoryCard(props: HistoryCardProps) {
         {keyArr.map((key, i) => <span key={key}>{`${keyHandler(key)}: ${valArr[i]}${(i + 1 > keyArr.length - 1) ? '' : '; '}`}</span>)}
       </div>
       <div className={styles.buttonBlock}>
-        <button type="submit" onClick={handleSearch}>open results</button>
-        <button type="submit" onClick={handleDelete}>delete</button>
+        <button type="submit" onClick={() => searchHistory(item)}>open results</button>
+        <button type="submit" onClick={() => refreshParent(item)}>delete</button>
       </div>
 
     </div>
@@ -48,6 +29,7 @@ function HistoryCard(props: HistoryCardProps) {
 HistoryCard.propTypes = {
   item: PropTypes.string.isRequired,
   refreshParent: PropTypes.func.isRequired,
+  searchHistory: PropTypes.func.isRequired,
 };
 
 type HistoryCardProps = InferProps<typeof HistoryCard.propTypes>;
